+++ src/app/lista-de-equipamentos/lista-de-equipamentos.page.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
 import { Equipamento } from '../classes/equipamento';
-import { EquipamentoService } from '../services/equipamento.service';
+import { EquipamentoService } from '../../services/equipamento.service';
 
 @Component({
   selector: 'app-lista-de-equipamentos',
