+++ src/app/lista-equipamentos-solicitados/lista-equipamentos-solicitados.page.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
-import { EquipamentoService } from '../equipamento.service';
-import { TipoEquipamentoService } from '../services/tipo-equipamento.service';
+import { EquipamentoService } from '../../services/equipamento.service';
+import { TipoEquipamentoService } from '../../services/tipo-equipamento.service';
 
 @Component({
   selector: 'app-lista-equipamentos-solicitados',
