+++ src/app/lista-materiais-entregues/lista-materiais-entregues.page.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
-import { MaterialService } from '../services/material.service';
+import { MaterialService } from '../../services/material.service';
 import { Material } from '../classes/material';
 
 @Component({
