+++ src/app/lista-sala/lista-sala.page.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
 import { Sala } from '../classes/sala';
-import { SalaService } from '../services/sala.service';
+import { SalaService } from '../../services/sala.service';
 
 @Component({
   selector: 'app-lista-sala',
