+++ src/app/lista-tipos-equipamento/lista-tipos-equipamento.page.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
-import { TipoEquipamentoService } from '../services/tipo-equipamento.service'
+import { TipoEquipamentoService } from '../../services/tipo-equipamento.service'
 
 @Component({
   selector: 'app-lista-tipos-equipamento',
