+++ src/services/entrega.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
-import { Entrega } from '../classes/entrega';
+import { Entrega } from '../app/classes/entrega';
 @Injectable({
   providedIn: 'root'
 })
