+++ src/services/equipamento.service.spec.ts
