+++ src/services/reserva-equipamento.service.spec.ts
