+++ src/services/tipo-equipamento.service.spec.ts
